Payments: Show closing transactions for trades
In the Payments view, there is a tab showing the individual transactions
that are aggregated in the tables and graphs of the other tabs. However, in
the mode that considers closed trades, this tab has been intentionally
empty.

There are reasons for and against this behaviour: On the one hand, a single
trade comprises multiple transactions (one or more opening and one closing),
some of which may be outside the time period being considered, and the
transactions (unlike for the other view such as fees, taxes) don’t
immediately show the profit of the trade. On the other hand, users have been
repeatedly confused that there were no transactions listed for trades, and
it makes verifying the aggregate results for trades hard.

Change this behaviour and from now on include (only) the closing transaction
of each trade. The closing transaction always is within the observed time
period; it gives the user a general idea of what has been included, and the
date allows looking up the trade in the separate Trades view easily, for
full details.