Remove duplicate transactions in the Payments view
The previous commit made closing transactions for trades show up in the
transactions tab in the Payments view. Because transactions from trades and
from other categories (fees, taxes, etc.) are added separately in the “sum”
mode, the same transaction may now appear twice if it both closes a trade
and contains a fee etc.; furthermore, transactions may be out of sequence,
because first all trade-related transactions are added to the list, then all
directly considered transactions.

To avoid this, collect all transactions in a set (removing duplicates)
before sorting the set’s content into a list.