Fixed wrong order of legend for yearly payments chart
SWTCharts sorts stacked series in reverse order in the legend. This
change reverses the order in which series are added in order to have
them correct for our use case.

Sometimes SWTChart showed the series in the correct order, because
due to the suspension of the chart during later updates, SWTChart did
not correctly recognized the series as stacked (and hence did not
reverse the order).

Issue: #3656
Issue: https://forum.portfolio-performance.info/t/reihenfolge-der-legende-in-ertragsansicht/25705
Issue: https://forum.portfolio-performance.info/t/legende-der-jahresertraege-bei-der-ersten-anwahl-falsch-sortiert/11444