Fixed the parsing of numbers with "non-standard" formatting Locales
The attribute implementation uses a pattern to check for valid numbers
before parsing the number. The pattern did not take local dependent
thousands and decimal separator into account.

With this change, the pattern is created based on the given locale.
To enable testing, the pattern is created when the converter is created.
Because converters are created only once per attribute, this should not
be a performance problem.

Issue: #3778
Co-authored-by: Lothar <github@kimmeringer.de>
Co-authored-by: Andreas Buchen <andreas.buchen@gmail.com>