Fixed usage of apostrophe in labels rendered using MessageFormat
If a label is used with MessageFormat, then single quotes in the text
must be represented by doubled single quotes '' throughout a String.
However, for translators it is not obvious in what context the label
is used.

When downloading from POEditor, the download tool will double the single
quotes if the string contains a curly brace (which indicates that it is
used in MessageFormat).

Issue: #3815