Added filter names and extension to save dialogs
On Windows and Linux, this change ensures that the picked filename has
the given file extension. If the user manually adds a file extension
in the input field, the default extension is appended.

On macOS, the users is asked if he wants to use a non-standard file
extension in case it has been changed.

Issue: #3826