Show a disabled version of the logo if there are no holdings
When going to the performance view (Performance -> Calculation) entries
of instruments that have no holdings are currently presented with the
generic logo of an inactive instrument regardless whether there is a
configured logo. With this change, we instead show a disabled image.

Issue: https://forum.portfolio-performance.info/t/logo-fehlt-in-performanceberechnung-bei-bestand-0/10321
Issue: #3977