Fix AppStream metainfo to avoid linter errors
The `appstreamcli` tool now validates the following rule about translations [1]:

> The description tag has to be translated by each <p> and <li> tags.

This means that the release notes description tags have to be structured as
follows:

```xml
<description>
  <ul>
    <li>Item in English.</li>
    <li>...</li>
    <li xml:lang="de">Item translated to German.</li>
    <li xml:lang="de">...</li>
  </ul>
</description>
```

Update the metainfo file to conform to this rule and fix the Flathub build.
For now, we are including this file in the downstream build recipe [2].

There is an accompanying pull request to the
portfolio-performance/portfolio-releng repository updating the release notes
generator to work with this new structure.

[1] https://docs.flathub.org/docs/for-app-authors/metainfo-guidelines/#metainfo-translations
[2]
https://github.com/flathub/info.portfolio_performance.PortfolioPerformance/commit/329148b6a7910cffec602f75e5f31dcb0eeaf8aa