Do not update dividend payments for retired instruments
Introduces the same behavior for updating dividend payments as is used
for updated historical prices. If the instrument is retired, then the
payments are only updated either if triggered for a specific instrument
or if triggered explicitly for all instruments. When opening the file,
the payments are not updated for retired instruments anymore.

Issue: #3951
Issue: #4001