Improve of the date recognition of the PDF-Importer (#4098)
The patter has been updated in the ExtractorUtils and the test cases.

The month recognition for non-German was changed from M to L in order to format the month designation according to the localized setting (locale) of the system.

It has also been added that the day can be a one-digit or two-digit number. The same applies to the month.