Separation of the numerical values in the security chart
There is the problem that too many numerical values are separated when displaying buy, sell, deposit... overlap as soon as the chart has a certain format.
This is how we separate the display of values from investment markers and dividend values.

Sorting label in menu structur
Rename SHOW_DATA_LABELS to SHOW_DATA_DIVESTMENT_INVESTMENT_LABEL