Allow '.' from numpad as French decimal separator
When entering a number of a French keyboard, the dot from the numpad
is usually automatically converted into the French decimal separator ','.

We do not have to change the NumberVerifyListener. It is not triggered
because the KeyListener is directly inserting the decimal. However, that
is not a problem because we know it is a valid character for numbers.

Issue: #4143
Issue: #3780
Issue: https://forum.portfolio-performance.info/t/decimal-separator-on-french-keyboard/28939
Signed-off-by: mierin12 <mattmailspoubelle+GIT@gmail.com>
[squashed commits; refactored code into utility class; rebased to master]
Signed-off-by: Andreas Buchen <andreas.buchen@gmail.com>