Security: Allow toString() be (in general) "ticker (name)"
For many people, ticker is more familiar than security name/title (this
is especially true for funds). toString() is used e.g. in dropdown boxes
for security selection. So, if a corresponding preference is enabled by
user, include ticker in the display string.

If ticker symbol is not available, or matches the security name, toString()
will return just "name", as before.