SecurityEvent: Move "source" field to the parent SecurityEvent class
Previously, "source" field was available only in the DividendEvent
subclass. But it makes sense to have it on the parent class, to
streamline integration with external tools to manage events. Thus,
"source" is a generic machine-readable field to record source and/or
"ownership" of the event information. A typical workflow would be:

1. An external tool TOOL1 creates an event and sets its source
attribute to "TOOL1" to designate this event is "owned" by the tool,
which it's free to modify or even delete later.
2. This attribute eventually gets rendered as <source> element within
<event> element in PortfolioPerformance XML.
3. PP preserves the value of the source attributes across load/save
cycle of the portfolio.
4. On the next run, TOOL1 can identify events originally create by
itself, and update them as needed. This is important, because event
information provided by external source if often tentative and
subject to change.

As mentioned above, the "source" field is intended to be machine-readable,
and as such, it's currently now shown anywhere in the PP UI. It might be
useful to address that at later time, if the need for that is found.

WARNING:
This change doesn't include steps which would be required to migrate
Protobuf binary format. It's expected that developers familiar with
this format would step in to help with the migration.