Fix "absolute deviation" in the "delta percentage indicator"
`node.getRoot()` leaded to an "absolute deviation" calculated in
relation to the total value of *all* (filtered) assets. The
filtering worked, but the "root" TaxonomyNode includes the
"Without Classification" class, which we don't want.

This commit fixes above issue, but goes further: multiple levels
of asset classes can now be rebalanced correctly using the
"delta percentage indicator", because only the current sub-class
is taken into account, as opposed to the classification root.

Issue:
https://forum.portfolio-performance.info/t/ohne-klassifizierungen-nicht-auswerten-filter-vergessen-ihre-einstellungen/655/9