Fix ASX trade calendar
The ASX observes Australia Day on the following Monday if January 26 happens
to fall on a weekend. This wasn’t represented in the PP trade calendar.

See the ASX trading calendar for 2025 and 2020 (Australia Day moved to
January 27) and 2019 (Australia Day moved to January 28) on their website
and on historic versions of it at the Internet Archive:
https://www.asx.com.au/markets/market-resources/trading-hours-calendar/cash-market-trading-hours/trading-calendar