AbstractSecurityListWidget: Show infopane on clicking a security
Previously, there was a code to update infopane to show security on which
user clicked in a dashboard list widget. But most people would probably
never know about this functionality, unless they happen to have infopane
enabled in the dashboard (which normally isn't). So, now explicitly show
infopane on click, so this useful functionality didn't go unnoticed.