ReportingPeriod: Adjust handling of today for recent Past, Future periods
Today cannot be resonably considered "past", so then it must go into the
Future class, to not have holes in the coverage. That also makes intuitive
sense on its own, as e.g. event an hour from now today is pretty much
future, and as we operate on the level of days, let the whole day be the
future yet.