Fixed resizing problem in instrument search dialog
Issue: if the dialog is resized smaller than the initial content, the
wizard is not propagating resize events to the content because of the
PageContainerFillLayout. Therefore an embedded scrollable control is
never triggered to show scroll bars.

Workaround: prevent the user from resizing the dialog smaller than the
initial size