Implement timezone field for JSON quote parser
The date format already provides a way to parse the time zone from the
string via appropriate patterns such as V or Z (see DateTimeFormatter
for available patterns), if it is part of the date representation chosen
by the server.
However, if the server doesn't provide an explicit timezone, yet the
timestamp is not in UTC, then there was no way to interpret it
appropriately. This commit allows the user to supply a fixed timezone
if they know the server uses it to represent dates.