Added a new minimal trade calendar
Used for checking if new prices are available at all. The current
'standard' calendar includes holidays such as christmas eve which are
not observed on all exchanges.

Introduced 'isSelectable' to avoid that the user can select the trade
calendar. We want to be able to change and remove this trade calendar
in the future.