Added detection of file type based on the content of the file only
Before, the detection of the file type (binary vs. compressed vs. XML)
was made based on the file name. However, on some platforms the app
does not get passed the file name, e.g., with the Google Drive
integration on Linux. This change - conservatively - adds an additional
check based on the content of the file if the detection by file name
failed.

Issue: https://forum.portfolio-performance.info/t/open-from-google-drive-xml-format-not-valid/37631