IBFlex: Improve handling of currency conversions
The existing exchange rate calculations are spread all over the
place and are not entirely consistent. For example, fxRateToBase
can only safely be used if the accountCurrency is known.

Improve this by using fxRateToBase correctly and by parsing
ConversionRate elements contained in the IB XML. This allows doing
cross currency conversions even if the accountCurrency is not known.