Offer the "Import" option really only on items with a warning.
Exactly what the source code comment says. This fixes the menu option
showing up for items having the new SKIP state.

Inside the code base the word "import" is used in ambiguous ways imo.
The meaning of the boolean flag "atLeastOneNotImported" is not clear at
last. It really is: atLeastOneWithoutAnImportOverrideAlthoughItCould
BeImportedAsItJustIsAWarning.

Issue: #5263